+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-10-16
 ### Added
-- Support subqery _metadata query(#4)
+- Support subqery \_metadata query(#4)
 - GraphiQL control flag - `playground`
 - Health check API `/.well-known/apollo/server-health`
 
@@ -34,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support GraphQL query style similar to subgraph.
 
-[Unreleased]: https://github.com/subquery/query-subgraph/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/subquery/query-subgraph/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/subquery/query-subgraph/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/subquery/query-subgraph/compare/v0.0.7...v0.1.0
 [0.0.7]: https://github.com/subquery/query-subgraph/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/subquery/query-subgraph/compare/v0.0.5...v0.0.6